Add preset flywheels as well as #name=value&name2=value2 parts to the URL
Presets and the hash part of the URL go hand-in-hand.  The URL is nice
because people can design a flywheel and "send" it to someone else fully
encoded into the URL.   Using the same syntax for the preset list makes
it possible to design a flywheel and copy/paste the hash part of the URL
to add it to the list.  Easy.