Reset prevFlywheelUnits when changing a preset.
If I don't do this, the a previous unit setting will be in effect and
when the flywheel design is changed from a preset, it will erroneously
try to do imperial/metric conversions.  Not good.

The preset contains the unit, so there can be both imperial and metric
preset flywheel designs.